# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EternaPy API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses grouped by learning path
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
...
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course detail with lessons
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course with lessons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
  /lessons/{lessonId}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get lesson detail with content
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"
  /lessons/{lessonId}/complete:
    post:
      operationId: completeLesson
      tags: [lessons]
      summary: Mark lesson as complete and earn XP
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonCompleteInput"
      responses:
        "200":
          description: Completion result with XP earned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/XpResult"
...
  /code/run:
    post:
      operationId: runCode
      tags: [code]
      summary: Execute code in a sandboxed environment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeRun"
      responses:
        "200":
          description: Code execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeRunResult"
...
  /progress/course/{courseId}:
    get:
      operationId: getCourseProgress
      tags: [progress]
      summary: Get user progress for a specific course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseProgress"
components:
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        language:
          type: ["string", "null"]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        estimatedHours:
          type: integer
        lessonCount:
          type: integer
        xpReward:
          type: integer
        isPro:
          type: boolean
        iconEmoji:
          type: string
        color:
          type: string
        tags:
          type: array
          items:
            type: string
        enrolledCount:
          type: integer
      required:
        - id
        - title
        - description
        - category
        - difficulty
        - estimatedHours
        - lessonCount
        - xpReward
        - isPro
        - iconEmoji
        - color
        - tags
        - enrolledCount
    CourseDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        longDescription:
          type: string
        category:
          type: string
        language:
          type: ["string", "null"]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        estimatedHours:
          type: integer
        xpReward:
          type: integer
        isPro:
          type: boolean
        iconEmoji:
          type: string
        color:
          type: string
        tags:
          type: array
          items:
            type: string
        enrolledCount:
          type: integer
        chapters:
          type: array
          items:
            $ref: "#/components/schemas/Chapter"
        userProgress:
          $ref: "#/components/schemas/CourseProgress"
      required:
        - id
        - title
        - description
        - longDescription
        - category
        - difficulty
        - estimatedHours
        - xpReward
        - isPro
        - iconEmoji
        - color
        - tags
        - enrolledCount
        - chapters
    Chapter:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        order:
          type: integer
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
      required:
        - id
        - title
        - order
        - lessons
    Lesson:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [reading, coding, quiz, video]
        order:
          type: integer
        xpReward:
          type: integer
        isCompleted:
          type: boolean
        durationMinutes:
          type: integer
      required:
        - id
        - title
        - type
        - order
        - xpReward
        - isCompleted
        - durationMinutes
    LessonDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [reading, coding, quiz, video]
        content:
          type: string
        xpReward:
          type: integer
        isCompleted:
          type: boolean
        starterCode:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        testCases:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/TestCase"
        nextLessonId:
          type: ["integer", "null"]
        prevLessonId:
          type: ["integer", "null"]
        courseId:
          type: integer
        courseTitle:
          type: string
      required:
        - id
        - title
        - type
        - content
        - xpReward
        - isCompleted
        - courseId
        - courseTitle
    TestCase:
      type: object
      properties:
        id:
          type: integer
        input:
          type: string
        expectedOutput:
          type: string
        description:
          type: string
      required:
        - id
        - input
        - expectedOutput
        - description
    LessonCompleteInput:
      type: object
      properties:
        timeSpentSeconds:
          type: integer
      required:
        - timeSpentSeconds
...
    XpResult:
      type: object
      properties:
        xpEarned:
          type: integer
        totalXp:
          type: integer
        newLevel:
          type: ["integer", "null"]
        newAchievements:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
        streakUpdated:
          type: boolean
      required:
        - xpEarned
        - totalXp
        - newAchievements
        - streakUpdated
...
    Achievement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        iconEmoji:
          type: string
        category:
          type: string
          enum: [beginner, intermediate, advanced, special]
        xpReward:
          type: integer
        requirement:
          type: string
      required:
        - id
        - title
        - description
        - iconEmoji
        - category
        - xpReward
        - requirement
...
    CodeRun:
      type: object
      properties:
        code:
          type: string
        language:
          type: string
        stdin:
          type: ["string", "null"]
      required:
        - code
        - language
    CodeRunResult:
      type: object
      properties:
        output:
          type: string
        error:
          type: ["string", "null"]
        exitCode:
          type: integer
        executionTimeMs:
          type: integer
      required:
        - output
        - exitCode
        - executionTimeMs
...
    CourseProgress:
      type: object
      properties:
        courseId:
          type: integer
        completedLessons:
          type: integer
        totalLessons:
          type: integer
        percentComplete:
          type: number
        lastLessonId:
          type: ["integer", "null"]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
      required:
        - courseId
        - completedLessons
        - totalLessons
        - percentComplete